requirements.txt
# Requirements for the Domo MCP Server
requests
python-dotenv
mcp